Using use col defs instead of transform data
